+++ src/yafowil/tests/__init__.py
@@ -17,6 +17,7 @@ def test_suite():
     from yafowil.tests import test_compound
     from yafowil.tests import test_controller
     from yafowil.tests import test_persistence
+    from yafowil.tests import test_resources
 
     suite = unittest.TestSuite()
 
@@ -25,6 +26,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_compound))
     suite.addTest(unittest.findTestCases(test_controller))
     suite.addTest(unittest.findTestCases(test_persistence))
+    suite.addTest(unittest.findTestCases(test_resources))
 
     return suite